#2642 Use a separate class for update and validate data
#2642 Flatrate Import: Support explicit enddate and terminated contracts
