+++ de.metas.contracts/src/main/java/de/metas/contracts/flatrate/impexp/FlatrateTermImportProcess.java
@@ -9,12 +9,9 @@ import org.adempiere.ad.trx.api.ITrx;
 import org.adempiere.impexp.AbstractImportProcess;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.lang.IMutable;
-import org.compiere.model.I_C_BPartner;
-import org.compiere.model.I_M_Product;
 import org.compiere.model.PO;
 import org.compiere.util.DB;
 
-import de.metas.contracts.model.I_C_Flatrate_Conditions;
 import de.metas.contracts.model.I_C_Flatrate_Term;
 import de.metas.contracts.model.I_I_Flatrate_Term;
 import de.metas.contracts.model.X_I_Flatrate_Term;
@@ -79,115 +76,7 @@ public class FlatrateTermImportProcess extends AbstractImportProcess<I_I_Flatrat
 	{
 		final String sqlImportWhereClause = COLUMNNAME_I_IsImported + "<>" + DB.TO_BOOLEAN(true)
 				+ "\n " + getWhereClause();
-
-		dbUpdateBPartnerIds(sqlImportWhereClause);
-		dbUpdateDropshipPartnerIds(sqlImportWhereClause);
-		dbUpdateC_Flatrate_Conditions_IDs(sqlImportWhereClause);
-		dbUpdateProductIdsByValue(sqlImportWhereClause);
-		dbUpdateProductIdsByName(sqlImportWhereClause);
-	}
-	
-	private void dbUpdateBPartnerIds(final String sqlImportWhereClause)
-	{
-		final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
-				+ " from " + I_C_BPartner.Table_Name + " bp "
-				+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_BPartnerValue
-				+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + "=(" + sqlSelectByValue + ")"
-				+ "\n WHERE " + sqlImportWhereClause
-				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL";
-
-		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
-		log.debug("Set C_BPartner_ID for {} records", no);
-
-		// Flag missing BPartners
-		markAsError("BPartner not found", I_I_Flatrate_Term.COLUMNNAME_C_BPartner_ID + " IS NULL"
-				+ "\n AND " + sqlImportWhereClause);
-	}
-
-	private void dbUpdateDropshipPartnerIds(final String sqlImportWhereClause)
-	{
-		final String sqlSelectByValue = "select MIN(bp." + I_C_BPartner.COLUMNNAME_C_BPartner_ID + ")"
-				+ " from " + I_C_BPartner.Table_Name + " bp "
-				+ " where bp." + I_C_BPartner.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value
-				+ " and bp." + I_C_BPartner.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + "=(" + sqlSelectByValue + ")"
-				+ "\n WHERE " + sqlImportWhereClause
-				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL";
-
-		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
-		log.debug("Set DropShip_BPartner_ID for {} records", no);
-
-		// Flag missing DropShip BPartners
-		markAsError("DropShip BPartner not found",
-				I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_ID + " IS NULL"
-						+ "\n AND " + I_I_Flatrate_Term.COLUMNNAME_DropShip_BPartner_Value + " IS NOT NULL" // only where the we have a key to match
-						+ "\n AND " + sqlImportWhereClause);
-	}
-
-	private void dbUpdateC_Flatrate_Conditions_IDs(final String sqlImportWhereClause)
-	{
-		final String sqlSelectByValue = "select MIN(fc." + I_C_Flatrate_Conditions.COLUMNNAME_C_Flatrate_Conditions_ID + ")"
-				+ " from " + I_C_Flatrate_Conditions.Table_Name + " fc "
-				+ " where fc." + I_C_Flatrate_Conditions.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_Value
-				+ " and fc." + I_C_Flatrate_Conditions.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + "=(" + sqlSelectByValue + ")"
-				+ "\n WHERE " + sqlImportWhereClause
-				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL";
-
-		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
-		log.debug("Set C_Flatrate_Conditions_ID for {} records", no);
-
-		// Flag missing flatrate conditions
-		markAsError("Flatrate conditions not found", I_I_Flatrate_Term.COLUMNNAME_C_Flatrate_Conditions_ID + " IS NULL"
-				+ "\n AND " + sqlImportWhereClause);
-	}
-
-	private void dbUpdateProductIdsByValue(final String sqlImportWhereClause)
-	{
-		final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
-				+ " from " + I_M_Product.Table_Name + " bp "
-				+ " where bp." + I_M_Product.COLUMNNAME_Value + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
-				+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
-				+ "\n WHERE " + sqlImportWhereClause
-				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
-
-		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
-		log.debug("Set M_Product_ID for {} records (by Value)", no);
-	}
-
-	private void dbUpdateProductIdsByName(final String sqlImportWhereClause)
-	{
-		final String sqlSelectByValue = "select MIN(bp." + I_M_Product.COLUMNNAME_M_Product_ID + ")"
-				+ " from " + I_M_Product.Table_Name + " bp "
-				+ " where bp." + I_M_Product.COLUMNNAME_Name + "=i." + I_I_Flatrate_Term.COLUMNNAME_ProductValue
-				+ " and bp." + I_M_Product.COLUMNNAME_AD_Client_ID + "=i." + I_I_Flatrate_Term.COLUMNNAME_AD_Client_ID;
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name + " i "
-				+ "\n SET " + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + "=(" + sqlSelectByValue + ")"
-				+ "\n WHERE " + sqlImportWhereClause
-				+ "\n AND i." + I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL";
-
-		final int no = DB.executeUpdateEx(sql, ITrx.TRXNAME_ThreadInherited);
-		log.debug("Set M_Product_ID for {} records (by Name)", no);
-
-		// Flag missing product
-		markAsError("Product not found", I_I_Flatrate_Term.COLUMNNAME_M_Product_ID + " IS NULL"
-				+ "\n AND " + sqlImportWhereClause);
-	}
-
-	private final void markAsError(final String errorMsg, final String sqlWhereClause)
-	{
-		final String sql = "UPDATE " + I_I_Flatrate_Term.Table_Name
-				+ "\n SET " + COLUMNNAME_I_IsImported + "=?, " + COLUMNNAME_I_ErrorMsg + "=" + COLUMNNAME_I_ErrorMsg + "||? "
-				+ "\n WHERE " + sqlWhereClause;
-		final Object[] sqlParams = new Object[] { X_I_Flatrate_Term.I_ISIMPORTED_ImportFailed, errorMsg + "; " };
-		DB.executeUpdateEx(sql, sqlParams, ITrx.TRXNAME_ThreadInherited);
-
+		FlatrateTermImportTableSqlUpdater.updateFlatrateTermImportTable(sqlImportWhereClause);
 	}
 
 	@Override
